Add a global asset store (#37)
This PR adds a global asset store in order to avoid asset duplication for different app versions. Each asset of an app will now be stored in the global store instead of local FS container.

New published applications will check along its MD5 sum if the asset exists in the global store. If not, the asset will be created. The CouchDB app version document will then update to link each asset filename with an object of the asset store.

Current applications will progressively update themselves. Each time an icon, screenshot or partnership_icon is requested and still located in the local store, it will me moved in the global one.