Use opts.Version instead of manifest.Version for tarball path (#43)
For dev and beta application versions, some webapps/konnectors use the
next version number as the "official" version number in the manifest (ie: 1.0.0
instead of 1.0.0-dev.abcdef). This was leading to a false generated
download URL as the registry was relying on this manifest version. It
should be fixed with this patch.