Fix opening a tarball
When the apps-registry try to open a file downloaded with the
application/octet-stream Content-Type for creating a new version of an
application, it doesn't know if it is a simple tarball or a .tar.gz.
It tries to open it as a .tar.gz, and if it fails, it tries to open it
as a simple tarball. But the later was failing because the first bytes
were read by trying to open it as .tar.gz. This commit fixes this issue
by making a seek call to come back to the start of the file.